adk_agent/requirements.txt
google-adk>=1.0.0
mcp[cli]>=1.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
